feat: implement escapeAfterConfirm
